backend/pyproject.toml
[project]
name = "visblow"
version = "0.1.0"
description = ""
authors = [
    {name = "Senyo197",email = "psenyo197@gmail.com"}
]
requires-python = ">=3.10"
dependencies = [
    "fastapi (>=0.134.0,<0.135.0)",
    "uvicorn[standard] (>=0.41.0,<0.42.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "pydantic-settings (>=2.13.1,<3.0.0)",
    "sqlalchemy[asyncio] (>=2.0.47,<3.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "alembic (>=1.18.4,<2.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "python-multipart (>=0.0.22,<0.0.23)",
    "redis (>=7.2.1,<8.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "structlog (>=25.5.0,<26.0.0)",
    "orjson (>=3.11.7,<4.0.0)",
    "celery (>=5.6.2,<6.0.0)",
    "psycopg[binary] (>=3.3.3,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
